backend/pyproject.toml
[project]
name = "lefil-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Kamigaku",email = "kamigaku@gmail.com"}
]
license = {text = "GPL-3.0"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "lxml (>=6.0.2,<7.0.0)",
    "readability-lxml (>=0.8.4.1,<0.9.0.0)",
    "trafilatura (>=2.0.0,<3.0.0)",
    "yake (>=0.7.3,<0.8.0)",
    "sumy (>=0.12.0,<0.13.0)",
    "nltk (>=3.9.3,<4.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "ollama (>=0.6.1,<0.7.0)",
    "structure @ file:///Code/LeFil/structure",
    "groq (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
